requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
python-dotenv==1.0.1

aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
attrs==26.1.0
bcrypt==4.3.0
blinker==1.9.0
branca==0.8.2
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.6
cssselect2==0.9.0
cycler==0.12.1
dnspython==2.8.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-mysql-connector==1.1.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
folium==0.20.0
fonttools==4.62.1
frozenlist==1.8.0
greenlet==3.2.4
gunicorn==25.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.2
matplotlib==3.10.8
multidict==6.7.1
mysql-connector-python==9.5.0
mysqlclient==2.2.8
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pdfkit==1.0.0
pillow==11.3.0
propcache==0.4.1
pycparser==3.0
pydyf==0.12.1
PyJWT==2.12.1
PyMySQL==1.1.2
pyparsing==3.3.2
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-http-client==3.3.7
pytz==2025.2
reportlab==4.4.10
requests==2.32.5
sendgrid==6.12.5
six==1.17.0
SQLAlchemy==2.0.43
tinycss2==1.5.1
tinyhtml5==2.1.0
twilio==9.10.4
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.0
waitress==3.0.2
weasyprint==68.1
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
xyzservices==2026.3.0
yarl==1.23.0
zopfli==0.4.1
